requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.7.1
    # via httpcore
async-timeout==4.0.2
    # via redis
attrs==23.1.0
    # via pytest
backoff==2.2.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-http
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
deprecated==1.2.14
    # via opentelemetry-api
environs==9.5.0
    # via
    #   gundi-client
    #   gundi-client-v2
exceptiongroup==1.1.2
    # via
    #   anyio
    #   pytest
google-api-core[grpc]==2.11.1
    # via
    #   google-cloud-pubsub
    #   google-cloud-trace
google-auth==2.22.0
    # via google-api-core
google-cloud-pubsub==2.18.2
    # via -r requirements.in
google-cloud-trace==1.11.2
    # via opentelemetry-exporter-gcp-trace
googleapis-common-protos[grpc]==1.60.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
    #   opentelemetry-exporter-otlp-proto-http
grpc-google-iam-v1==0.12.6
    # via google-cloud-pubsub
grpcio==1.56.2
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.56.2
    # via
    #   google-api-core
    #   google-cloud-pubsub
gundi-client==1.0.1
    # via -r requirements.in
gundi-client-v2==2.1.1
    # via -r requirements.in
gundi-core==1.2.2
    # via
    #   -r requirements.in
    #   gundi-client
    #   gundi-client-v2
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via
    #   gundi-client
    #   gundi-client-v2
    #   respx
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
marshmallow==3.20.1
    # via environs
opentelemetry-api==1.14.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-propagator-gcp
    #   opentelemetry-sdk
opentelemetry-exporter-gcp-trace==1.3.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-http==1.15.0
    # via -r requirements.in
opentelemetry-instrumentation==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
opentelemetry-instrumentation-aiohttp-client==0.35b0
    # via -r requirements.in
opentelemetry-instrumentation-httpx==0.35b0
    # via -r requirements.in
opentelemetry-instrumentation-requests==0.35b0
    # via -r requirements.in
opentelemetry-propagator-gcp==1.5.0
    # via -r requirements.in
opentelemetry-proto==1.15.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.14.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-sdk
opentelemetry-util-http==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-requests
packaging==23.1
    # via
    #   marshmallow
    #   pytest
pluggy==1.2.0
    # via pytest
proto-plus==1.22.3
    # via
    #   google-cloud-pubsub
    #   google-cloud-trace
protobuf==4.23.4
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   opentelemetry-proto
    #   proto-plus
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==1.10.4
    # via
    #   -r requirements.in
    #   gundi-client
    #   gundi-client-v2
    #   gundi-core
pytest==7.2.1
    # via
    #   -r requirements.in
    #   pytest-mock
pytest-mock==3.10.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via environs
redis==4.6.0
    # via walrus
requests==2.31.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-http
respx==0.20.2
    # via
    #   gundi-client
    #   gundi-client-v2
rsa==4.9
    # via google-auth
six==1.16.0
    # via google-auth
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tomli==2.0.1
    # via pytest
typing-extensions==4.7.1
    # via
    #   opentelemetry-sdk
    #   pydantic
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
walrus==0.9.2
    # via -r requirements.in
wrapt==1.15.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client

# The following packages are considered to be unsafe in a requirements file:
# setuptools
